% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Assets}
\alias{Assets}
\title{Assets}
\format{A data frame with 4 columns}
\usage{
Assets
}
\description{
Data set containg three numerical columns, value, duration, and yield of consolidated balance sheet assets.
   The data set is  used in the examples provided in the vignette.

\itemize{
  \item Deutsche. Descritption of the assets in the asset portfolio
  \item Asset. Values of the assets from the assetportfolio
  \item Duration. The durations of the assets from the asset portfolio
  \item Yield. The yields of the assets from the asset portfolio
}
}
\details{
Consolidated balancesheet's asset information
}
\references{
Deutsche Bank Annual Report 2018, Consolidated Balance Sheet
}
\keyword{datasets}
