% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change}
\alias{change}
\title{Computes the aggregate change in assets or liabilities, due to a respective predicted change in interest rate level}
\usage{
change(x, y, z, r)
}
\arguments{
\item{x}{Vector containing asset or liabilities positions in asset or liability portfolio}

\item{y}{Vector containing duration of asset or liabilities in asset or liability portfolio}

\item{z}{Vector containing the yield of the asset or liabilities in asset or liability portfolio}

\item{r}{The predicted change in ineterst rate level}
}
\value{
aggregate change in assets or liabilities
}
\description{
Takes the positions of assets or liabilities together with the respective durations and yields
    and computes the expected overall change in the value of assets or liabilities resulting from the predicted change in interest rate level
}
\examples{
change(c(150,350,600), c(0.25, 2.5, 0.75), c(0.01, 0.035, 0.65), 0.0015)
}
