% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaE.R
\name{deltaE}
\alias{deltaE}
\title{Computes the change in equity}
\usage{
deltaE(x, y, z, w, r)
}
\arguments{
\item{x}{Vector containing asset positions in asset portfolio}

\item{y}{Vector containing durations of assets from the asset portfolio}

\item{z}{Vector containing liability positions from the liabilities portfolio}

\item{w}{Vector containing duration of liabilities from the liabilities portfolio}

\item{r}{The predicted change in interest rate level across both assets and liabilities}
}
\value{
leverage adjusted duration gap
}
\description{
Takes the duration of assets and the duration of liabilities,
    applies them to the leverage adjusted duration gap computation, ultimately applies
    the multiplication of negative ladg and total assets and the predicted change in interest rate level
}
\examples{
deltaE(c(150,350,600), c(0.25, 2.5, 0.75), c(200, 375, 120), c(0.1, 2, 0.75), -0.003)
}
