% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differenceE.R
\name{differenceE}
\alias{differenceE}
\title{Computes the change in equity}
\usage{
differenceE(x, y, z, r1, h, i, j, r2)
}
\arguments{
\item{x}{Vector containing asset positions from asset portfolio}

\item{y}{Vector containing duration of assets from the asset portfolio}

\item{z}{Vector containing asset yields from the asset portfolio}

\item{r1}{The predicted change in interest rate level for assets}

\item{h}{Vector containing values of liabilities from the liabilities portfolio}

\item{i}{Vector containing duration of liabilities from the liabilities portfolio}

\item{j}{Vector containing the yields of liabilities from the liabilities portfolio}

\item{r2}{The predicted change in interest rate level for liabilities}
}
\value{
The change in equity
}
\description{
Takes the aggregare changes in assets and liabilities, subtracting liabilities from assets, returning the change in equity.
    Note: used when the predicted change in interest level for assets and liabilities asre different. (spread > 0)
}
\examples{
differenceE(c(150,350,600), c(0.25,2.5,0.75), c(3,0.45,0.6),0.0015, c(200,375,120), c(0.1,2,0.75),c(2,1.2,0.35),0.0025)
}
