% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\name{duration}
\alias{duration}
\title{Computes the duration of assets or liabilities}
\usage{
duration(x, y)
}
\arguments{
\item{x}{Vector containing asset or liabilities positions in liability or asset portfolio}

\item{y}{Vector containing duration of asset or liabilities in liability or asset portfolio}
}
\value{
aggregate duration of assets or liabilities
}
\description{
Takes the positions of assets or liabilities, computes their porportions relative to the total liability portfolio or asset portfolio,
    and returns the aggregate duration.
}
\examples{
duration(c(150,350,600), c(0.25, 2.5, 0.75))
}
