% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladg.R
\name{ladg}
\alias{ladg}
\title{Computes the leverage adjusted duration gap}
\usage{
ladg(x, y, z, w)
}
\arguments{
\item{x}{Vector containing asset positions from the asset portfolio}

\item{y}{Vector containing duration of assets from the asset portfolio}

\item{z}{Vector containing liability positions from the liabilities portfolio}

\item{w}{Vector containing duration of liabilities from the liabilities portfolio}
}
\value{
leverage adjusted duration gap
}
\description{
Takes the duration of assets and the scaled duration of liabilities,
    subtracting the scaled duration of liabilities from the duration of assets
}
\examples{
change(c(150,350,600), c(0.25, 2.5, 0.75), c(200, 375, 120), c(0.1, 2, 0.75))
}
