% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Liabilities}
\alias{Liabilities}
\title{Liabilities}
\format{A data frame with 4 columns}
\usage{
Liabilities
}
\description{
A data set containing the consolidated balance sheet liabilities values, liabilities durations, and liabilities yields.

 \itemize{
  \item Deutsche. Descritption of the liabilities
  \item Liabilities. Value of the liabilities
  \item duration. The duration of the liabilities
  \item yield. The yield of the liabilities
}
}
\details{
Consolidated balance sheet's liabilities information
}
\references{
Deutsche Bank Annual Report 2018, Consolidated Balance Sheet
}
\keyword{datasets}
